# docs/api/api_v3/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Koudaisai Portal API v3",
    "version": "0.1.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "description": "Get all users.",
        "operationId": "get_users",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "description": "Create a new user (id is generated server-side); returns the activation token.",
        "operationId": "post_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created; returns the activation token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "m_address already in use"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "description": "Get the currently authenticated user.",
        "operationId": "get_user_me",
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "description": "Get a user by id.",
        "operationId": "get_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "description": "Delete a user by id.",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "description": "Edit a user's name by id (m_address is changed via its own endpoint).",
        "operationId": "patch_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{id}/m_address": {
      "post": {
        "tags": [
          "users"
        ],
        "description": "Change a user's m_address and re-issue an activation token.",
        "operationId": "post_user_m_address",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MAddressUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "m_address changed; returns a new activation token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MAddressUpdated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{id}/read-notifications": {
      "get": {
        "tags": [
          "users"
        ],
        "description": "Get the notifications the user has marked as read.",
        "operationId": "get_user_read_notifications",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationRead"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MAddressUpdate": {
        "type": "object",
        "description": "`POST /users/{id}/m_address` のリクエストボディ。",
        "required": [
          "m_address"
        ],
        "properties": {
          "m_address": {
            "type": "string"
          }
        }
      },
      "MAddressUpdated": {
        "type": "object",
        "description": "m アドレス変更時に再発行される activation token を含むレスポンス。",
        "required": [
          "activation_token"
        ],
        "properties": {
          "activation_token": {
            "type": "string"
          }
        }
      },
...
      "NotificationRead": {
        "allOf": [
          {
            "$ref": "#/components/schemas/NotificationType"
          },
          {
            "type": "object",
            "required": [
              "id",
              "created_at",
              "updated_at",
              "targets"
            ],
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "created_by": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "uuid"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "targets": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "NotificationType": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "title",
              "content",
              "type"
            ],
            "properties": {
              "content": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "markdown"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "approval_request_id",
              "type"
            ],
            "properties": {
              "approval_request_id": {
                "type": "string",
                "format": "uuid"
              },
              "type": {
                "type": "string",
                "enum": [
                  "approval_request"
                ]
              }
            }
          }
        ]
      },
...
      "UserCreate": {
        "type": "object",
        "required": [
          "name",
          "m_address"
        ],
        "properties": {
          "m_address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UserCreated": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserRead"
          },
          {
            "type": "object",
            "required": [
              "activation_token"
            ],
            "properties": {
              "activation_token": {
                "type": "string"
              }
            }
          }
        ],
        "description": "`POST /users` でユーザーを新規作成したときのレスポンス。\n作成された(`Registered`)ユーザーが初回ログインで有効化するための\nactivation token を含む（このトークンは作成時にのみ返却される）。"
      },
      "UserRead": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserReadStatus"
          },
          {
            "type": "object",
            "required": [
              "id",
              "created_at",
              "updated_at",
              "name",
              "m_address"
            ],
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "m_address": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "UserReadStatus": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "status"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "registered"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "status"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "active"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "deactivated_at",
              "reason",
              "status"
            ],
            "properties": {
              "deactivated_at": {
                "type": "string",
                "format": "date-time"
              },
              "reason": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "deactivated"
                ]
              }
            }
          }
        ],
        "description": "`status` フィールドを判別子に持つ internally-tagged enum。\nflatten 時に `{\"status\": \"registered\", ..}` のように展開される\n(兄弟の `FormType` / `ApprovalRequestStatus` 等と同じ表現)。"
      },
      "UserUpdate": {
        "type": "object",
        "description": "`PATCH /users/{id}` のリクエストボディ。氏名のみ変更可能\n(m アドレス変更は専用エンドポイント `POST /users/{id}/m_address`)。",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    }
  }
}